Abstract away acceptNftSellOffer
